same filename of rdjk and rdjK caused problems on mac (case insensitive)
